Load dropzone component asynchronously, just after main app loading
